Raise the correction-pulse ceiling to 100 ms
The coil measured 48R; through R6 (10R) that's ~0.21A - about 0.4W in
R6 and ~2.1W in the coil, well within what COILTEST already proved
both can shed continuously for 5 whole seconds. No thermal reason left
to cap a single pulse at 50 ms.

Widening DRIVE_MAX_PULSE_US to 100000 exposed a real bug: cfg.pulse_us
was a uint16_t (max 65535), so any width above that would have
silently wrapped instead of clamping to the new ceiling. Widened it to
uint32_t, matching pulse_advance_us/pulse_retard_us's own already-wider
handling, and bumped CONFIG_VERSION since the layout changed.

+++ Code/synchronizer/src/cli.c
@@ -464,7 +464,7 @@ static int coiltest_cmd(int args, tinycl_parameter *tp, void *v)
 static int pw_cmd(int args, tinycl_parameter *tp, void *v)
 {
   (void)args; (void)v;
-  cfg.pulse_us = (uint16_t)tp[0].ti.i;
+  cfg.pulse_us = (uint32_t)tp[0].ti.i;
   if (cfg.pulse_us > DRIVE_MAX_PULSE_US) cfg.pulse_us = DRIVE_MAX_PULSE_US;
   printf("pulse width %u us\r\n", cfg.pulse_us);
   return 1;
